✨ (MessageInput): refactor MessageInput component to use a new Input subcomponent for better organization and readability ♻️ (nice.tsx): rename Send icon component to Nice for improved clarity 📝 (MessageInput.stories.tsx): simplify story structure by removing unnecessary Box and Toolbar components for cleaner presentation
✨ (Toolbar.tsx): introduce a new Toolbar component to enhance message input functionality
✨ (Text.tsx): add a new Text component for displaying message content
