+++ src/Molecules/Message/MessageItem/components/Text/Text.tsx
@@ -1,0 +1,3 @@
+export const Text = () => {
+  return <div>Text</div>;
+};
